Fix the display of the timestamp
The timestamp was displayed on the x-axis and therefore hidden.
This change displays the label at a fixed position relative to the
figure (in the top left corner).
Depending on the aspect ratio the text might overflow the box.
This shouldn't be a problem in generated videos as they have an
according aspect ratio.