Refactor code to load nt,et,r_com from within plotSat.m
Included r_com vector to move the spacecrafts center of mass onto the
origin of the spacecraft coordinate system.
Splitting this into the different function increases code readability
but might be a poor choice performance wise, as it has to load the
files nt.txt/et.txt/ct.txt and the value of r_com in every iteration.

Also fixed a minor bug while plotting in debug mode by enabling hold on,
to stop overwriting the plot with each newly plotted surface.