Add new view modes
Adding new view modes that replace and extend the old modes.
This commit gives the user the ability to specify a custom camera
perspective without having to use default presets.

The new modes are:
earthCentered = focus around earth
satCentered = focus around sat
satFixed = focus around sat and move with the satellites coordinate

All of them have the added ability to rotate the camera with cRoll
and to orbit around the center point (earth or satellite) with cOrb.

The old view modes are still in place.