Fix time stamp breaking into two lines in 720p
This fixes a bug that scaled the time box depending on the window width.
Since Matlab annotation boxes only take relative widths as an argument
this change calculates the relative width. It does so by using the
reciprocal of the pixel width times the window width.

A problem would still occur if the textstring was longer as the textbox
pixel widths is still hard coded. Since the string length should not
change (months are abbreviated to 3 chars)