Fix the hopping Bug when elements get off screen
The Bug had the entire scene "hop" when something moved out off the
screen range.
The way to disable the coordinate Box is set differently for every
view mode depending on the focus.
It still changes the size in satCentered mode a little. But the movement
is smoothed out through the `axis image` command.