Add function to display the ground track
This is a new feature requested by the user base.
This feature adds a queue in which the (onto the sphere projected)
positional data of the spacecraft is stored. It is then plotted.
The length of the queue defines how far the ground track is displayed
into the past and can be set via the groundTrackLenght setting.