Update ground track to rotate with earth
This extends the ground track function to rotate with the earth by
rotating the groundTrack matrix with the rot_ang vector.

The groundTrack matrix is transformed by the rotation matrix generated
with the rot_angle, the satellites position is then added to the matrix
before the matrix is transformed by the inverse rotation matrix.

The changes lead to a slight performance decrease.