Remove plotting within the for loop
This commit increases the performance by a lot and also works towards
removing the jumping bug.
By pulling out the plotting from the for loop the objects only get
plotted once and are only positioned and rotated within the for loop.
Most testing has to start over probably and there might be a significant
amount of new bugs.